# backend/src/main/java/eu/happybit/poller/domain/ApiError.java
package eu.happybit.poller.domain;

import java.time.LocalDateTime;
import java.util.List;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class ApiError {
    private int status;
    private LocalDateTime timestamp;
    private List<String> errors;
}